refactor(tui): extract the render layer into tui/render.rs
Move all render_* widgets and their pure label/format/diff helpers into a
child module. render.rs uses `use super::*` to reach AppState/enums and the
helpers left in mod.rs (a child sees the parent's private items), so only the
render entry points it exposes are marked pub(super). Pure refactor, no
behaviour change. Part of #73.
